+++ NEWS.md
@@ -1,6 +1,6 @@
 # ecotraj 1.1.0
 * New functions is.metric() and is.synchronous().
-* Update of function trajectoryVariability().
+* Update of function trajectoryInternalVariation() (former trajectoryVariability()).
 * New function variationDecomposition() for synchronous trajectories.
 * New function interpolateTrajectories() for interpolation to given times.
 * Function trajectoryConvergence() modified to allow a global convergence test (type = "multiple").